refactor: Split code
